// employee-managementUI/src/app/models/userModel.ts
export interface User{
    id: number;
    firstName: string;
    lastName: string;
    birthDate: Date;
    username: string;
    passwordHash: string;
    email: string;
    phoneNumber: string;
    address: string;
    city: string;
    district: string;
    division: string;
    country: string;
    role: string;
}